Fixed #15
SubGraph doesn't always pass control to next node

Exit nodes do not exit themselves. Instead they hang as 'active'. A node that is active cannot be activated again, so when the subgraph finishes execution the first time, the Exit node correctly passes control to the parent graph in its OnEnter override. The second time the sub graph runs, the Exit node is already active, so its OnEnter is ignored, and the SubGraph node stalls.

Fixed by simply adding Exit() to the end of OnEnter in the Exit node.